added data on the top asset section
