+++ components/traders_comp/dashboard/TopAssets.jsx
@@ -3,8 +3,123 @@ import Link from "next/link";
 import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/css";
 import "swiper/css/pagination";
+import useAllCryptoCoins from "@/hooks/useAllCryptoCoins";
+import useAllFlatCoins from "@/hooks/useAllFlatCoins";
+import { useEffect, useState } from "react";
 
 const TopAssets = () => {
+  const [cryptoCurrency, setCryptoCurrency] = useState([]);
+  const [flatCurrency, setFlatCurrency] = useState([]);
+
+  // get currency data from hook
+  const { allCryptoCoins, cryptoRefetch } = useAllCryptoCoins();
+  const { allFlatCoins, flatRefetch } = useAllFlatCoins();
+
+  // data without real time price
+  useEffect(() => {
+    if (allCryptoCoins.length > 0 && allFlatCoins.length > 0) {
+      setCryptoCurrency(allCryptoCoins);
+      setFlatCurrency(allFlatCoins);
+    }
+  }, [allCryptoCoins, allFlatCoins]);
+
+  // createData function for adding real time prices
+  const createData = (
+    _id,
+    name,
+    key,
+    price,
+    type,
+    changePrice,
+    highPrice,
+    lowPrice,
+    icon
+  ) => ({
+    _id,
+    name,
+    key,
+    price,
+    type,
+    changePrice,
+    highPrice,
+    lowPrice,
+    icon,
+  });
+
+  // fetch real time crypto data from binance api and create data
+  useEffect(() => {
+    const socket = new WebSocket(
+      "wss://stream.binance.com:9443/ws/!ticker@arr"
+    );
+
+    socket.addEventListener("message", (event) => {
+      const data = JSON.parse(event.data);
+      if (cryptoCurrency.length > 0) {
+        const updatedAssets = cryptoCurrency.map((asset) => {
+          const ticker = data.find((item) => item.s === asset.key);
+          if (ticker) {
+            return createData(
+              asset._id,
+              asset.name,
+              asset.key,
+              parseFloat(ticker.c).toFixed(3),
+              asset.type,
+              parseFloat(ticker.p).toFixed(3),
+              parseFloat(ticker.h).toFixed(2),
+              parseFloat(ticker.l).toFixed(2),
+              asset.icon
+            );
+          }
+          return asset;
+        });
+        setCryptoCurrency(updatedAssets);
+      }
+    });
+    return () => socket.close();
+  }, [cryptoCurrency]);
+
+  // createFlatCurrencyData function for adding current price
+  const createFlatCurrencyData = (_id, name, key, type, price, icon) => ({
+    _id,
+    name,
+    key,
+    type,
+    price,
+    icon,
+  });
+
+  // fetch current flat coin prices from exchange rate api and create data
+  useEffect(() => {
+    const fetchCurrencyRates = async () => {
+      try {
+        if (flatCurrency.length > 0) {
+          const response = await axios.get(
+            "https://api.exchangerate-api.com/v4/latest/USD"
+          );
+          // Access the data property of the response to get the currency rates
+          const data = response.data.rates;
+          const updatedAssets = flatCurrency.map((cur) => {
+            const currencyKey = cur.key;
+            // console.log(currencyKey)
+            return createFlatCurrencyData(
+              cur._id,
+              cur.name,
+              cur.key,
+              cur.type,
+              data[currencyKey],
+              cur.icon
+            );
+          });
+          setFlatCurrency(updatedAssets);
+        }
+      } catch (error) {
+        console.error("Error fetching currency rates:", error);
+      }
+    };
+
+    fetchCurrencyRates();
+  }, [flatCurrency]);
+
   return (
     <div className="xl:col-span-6 2xl:col-span-12 3xl:col-span-7 h-80">
       <div className="flex items-center justify-between">
