+++ components/traders_comp/dashboard/quick-trade/quick-buy/BuyLimit.jsx
@@ -3,7 +3,12 @@ import { MdCheckBox, MdCheckBoxOutlineBlank } from "react-icons/md";
 import CryptoBuy from "./CryptoBuy";
 import CurrencyBuy from "./CurrencyBuy";
 
-const BuyLimit = ({ cryptoCurrency, flatCurrency, cryptoRefetch, flatRefetch }) => {
+const BuyLimit = ({
+  cryptoCurrency,
+  flatCurrency,
+  cryptoRefetch,
+  flatRefetch,
+}) => {
   const [isOpenSetLimit, setIsOpenSetLimit] = useState(false);
   const [limitAmount, setLimitAmount] = useState(0);
   const [isLimitAmountError, setIsLimitAmountError] = useState(false);
@@ -99,7 +104,14 @@ const BuyLimit = ({ cryptoCurrency, flatCurrency, cryptoRefetch, flatRefetch }) 
         </div>
       </div>
 
-      {isCryptoSelected ? <CryptoBuy cryptoCurrency={cryptoCurrency}  cryptoRefetch={cryptoRefetch}  /> : <CurrencyBuy flatCurrency={flatCurrency} flatRefetch={flatRefetch}/>}
+      {isCryptoSelected ? (
+        <CryptoBuy
+          cryptoCurrency={cryptoCurrency}
+          cryptoRefetch={cryptoRefetch}
+        />
+      ) : (
+        <CurrencyBuy flatCurrency={flatCurrency} flatRefetch={flatRefetch} />
+      )}
 
       <div className="bg-darkBG flex items-center justify-between gap-5 px-5 py-[10px] rounded-full my-5">
         <p className="text-sm font-medium">Trade amount</p>
